pyproject.toml
[tool.poetry]
name = "recipe-app-api"
version = "0.1.0"
description = "Recipe app api source code."
authors = ["Carlos Alberto Giralt Torriente <cagiraltt@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
flake8 = "^6.0.0"
black = "^22.12.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
    /(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | __pycache__
  | migrations
  | build
  | dist
  | docs
  | node_modules
    )/
  '''
